Update state folder structure to mimic the state subtree
Update state folder structure as per Calypso’s data recommendations -
https://goo.gl/2mLzdB

Rename schema to match name of state subtree that it's describing

Remove unused selector
