+++ client/extensions/wp-super-cache/state/settings/schema.js
@@ -1,4 +1,4 @@
-export const settingsSchema = {
+export const itemsSchema = {
 	type: 'object',
 	additionalProperties: false,
 	patternProperties: {
